requirements.txt
hermes_python>=0.1
future
paho-mqtt
